Pushing changes to introduce simplified utils for ithriv setup and restructure models for better handling and eliminating redundancy
