+++ backend/app/resources/CategoryEndoint.py
@@ -4,7 +4,7 @@ from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import joinedload
 
 from app import db, RestException
-from app.model.category import Category
+from app.models import Category
 from app.resources.schema import CategorySchema, ParentCategorySchema
 from app.resources.Auth import login_optional
