+++ backend/app/resources/Tracking.py
@@ -3,8 +3,8 @@
 import datetime
 
 from app import sso, app, RestException, db, auth
-from app.model.email_log import EmailLog
-from app.model.user import User
+from app.models import EmailLog
+from app.models import User
 from flask import jsonify, redirect, g, request, Blueprint, send_file
 
 tracking_blueprint = Blueprint('track', __name__, url_prefix='/api/track')
