+++ backend/requirements.txt
@@ -1,17 +1,64 @@
-Flask==0.12.2
-Flask-RESTful==0.3.6
-marshmallow==2.15.0
-Flask-Migrate==2.1.1
-psycopg2==2.7.5
-Flask-Marshmallow==0.7.0
-marshmallow_sqlalchemy==0.13.2
+alembic==1.0.8
+aniso8601==6.0.0
+asn1crypto==0.24.0
+attrs==19.1.0
+Automat==0.7.0
+bcrypt==3.1.6
+blinker==1.4
+boto3==1.7.40
+botocore==1.10.84
+cffi==1.12.2
+Click==7.0
+constantly==15.1.0
+cryptography==2.6.1
+cssselect==1.0.3
+docutils==0.14
 elasticsearch==6.2.0
 elasticsearch-dsl==6.1.0
-scrapy==1.5.0
-flask-cors==3.0.3
-boto3==1.7.40
-python-magic==0.4.13
+Flask==0.12.2
+Flask-Bcrypt==0.7.1
+Flask-Cors==3.0.3
+Flask-HTTPAuth==3.2.4
+flask-marshmallow==0.7.0
+Flask-Migrate==2.1.1
+Flask-RESTful==0.3.6
+Flask-SQLAlchemy==2.3.2
+Flask-SSO==0.4.0
+hyperlink==18.0.0
+idna==2.8
+incremental==17.5.0
+ipaddress==1.0.22
+itsdangerous==1.1.0
+Jinja2==2.10
+jmespath==0.9.4
+lxml==4.3.2
+Mako==1.0.7
+MarkupSafe==1.1.1
+marshmallow==2.15.0
+marshmallow-sqlalchemy==0.13.2
+parsel==1.5.1
+psycopg2==2.7.5
+psycopg2-binary==2.7.7
+pyasn1==0.4.5
+pyasn1-modules==0.2.4
+pycparser==2.19
+PyDispatcher==2.0.5
+PyHamcrest==1.9.0
 PyJWT==1.6.3
-flask-sso==0.4.0
-flask-httpauth==3.2.4
-flask-bcrypt==0.7.1
+pyOpenSSL==19.0.0
+python-dateutil==2.8.0
+python-editor==1.0.4
+python-magic==0.4.13
+pytz==2018.9
+queuelib==1.5.0
+s3transfer==0.1.13
+Scrapy==1.5.0
+service-identity==18.1.0
+six==1.12.0
+SQLAlchemy==1.3.1
+SQLAlchemy-Utils==0.33.11
+Twisted==18.9.0
+urllib3==1.22
+w3lib==1.20.0
+Werkzeug==0.14.1
+zope.interface==4.6.0
